% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitr.utils.r
\name{add.SIT.info2post}
\alias{add.SIT.info2post}
\title{Add info to the post. Custom version for SIT blog.}
\usage{
add.SIT.info2post(post.filename, out.filename,
  fig.path = post.figure.path(post.filename), add.source.post.link = F)
}
\arguments{
\item{post.filename}{R script post filename}

\item{out.filename}{output filename}

\item{fig.path}{figure folder, \strong{defaults to post.figure.path(post.filename)}}

\item{add.source.post.link}{flag to indicate to include source post link, \strong{defaults to FALSE}}
}
\value{
output filename
}
\description{
Add info to the post. Custom version for SIT blog.
}
\examples{
\dontrun{
filename = 'post.test.r'
temp.filename = 'post.r'
make.rmd(add.SIT.info2post(filename, temp.filename))
}
}

